+++ front/src/modules/people/components/__stories__/PeopleFullNameEditableField.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { mockedPeopleData } from '~/testing/mock-data/people';
 import { getRenderWrapperForComponent } from '~/testing/renderWrappers';
 
-import { PeopleFullNameEditableField } from '../PeopleFullNameEditableField';
+import { PeopleFullNameEditableField } from '../../editable-field/components/PeopleFullNameEditableField';
 
 const meta: Meta<typeof PeopleFullNameEditableField> = {
   title: 'Modules/People/EditableFields/PeopleFullNameEditableField',
